Kops - extract templates and helper fns into their own files
